mensajes/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
logging